requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
astroid==2.15.6
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
black==23.7.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
dill==0.3.6
exceptiongroup==1.1.2
fastapi==0.100.0
frozenlist==1.4.0
h11==0.14.0
httptools==0.6.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mariadb==1.0.11
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.27.8
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
pydantic==2.0.3
pydantic_core==2.3.0
pylint==2.17.4
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
tomli==2.0.1
tomlkit==0.11.8
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.15.0
yarl==1.9.2
